refactor(wow-core): extract AbstractWaitStrategy and simplify WaitingFor implementations
- Extract common wait strategy logic into AbstractWaitStrategy class
- Simplify WaitingFor interface and remove duplicate code
- Update WaitingForAfterProcessed and WaitingForStage to extend AbstractWaitStrategy
- Remove redundant methods and properties from WaitingFor implementations
